Update delete day offf

+++ lib/src/presentation/pages/service_canlendar/service_calendar_page_controller.dart
@@ -136,6 +136,23 @@ class ServiceCalendarPageController extends GetxController {
 
   Future<void> deleteDayOffInfo(DayOffInfoModel info) async {
     try {
+      final isConfirm = await Get.dialog(
+        AlertDialog(
+          title: const Text('Xác nhận'),
+          content: const Text('Bạn có chắc chắn muốn xóa ngày nghỉ này?'),
+          actions: [
+            TextButton(
+              onPressed: () => Get.back(result: false),
+              child: const Text('Hủy'),
+            ),
+            TextButton(
+              onPressed: () => Get.back(result: true),
+              child: const Text('Xóa'),
+            ),
+          ],
+        ),
+      );
+      if (isConfirm != true) return;
       await _partnerDayOffService.deletePartnerDayOff(info.id);
       SnackBarHelper.show(message: 'Xóa ngày nghỉ thành công');
       _loadDayOffInMonth();
